+++ macos/mobile_scanner/Sources/mobile_scanner/MobileScannerPlugin.swift
@@ -131,7 +131,9 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
                         
                         if error != nil {
                             DispatchQueue.main.async {
-                                self?.sink?(FlutterError(code: "MobileScanner", message: error?.localizedDescription, details: nil))
+                                self?.sink?(FlutterError(
+                                    code: MobileScannerErrorCodes.BARCODE_ERROR,
+                                    message: error?.localizedDescription, details: nil))
                             }
                             return
                         }
@@ -180,9 +182,11 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
                     }
 
                     try imageRequestHandler.perform([barcodeRequest])
-                } catch let e {
+                } catch let error {
                     DispatchQueue.main.async {
-                        self?.sink?(FlutterError(code: "MobileScanner", message: e.localizedDescription, details: nil))
+                        self?.sink?(FlutterError(
+                            code: MobileScannerErrorCodes.BARCODE_ERROR,
+                            message: error.localizedDescription, details: nil))
                     }
                 }
             }
@@ -262,8 +266,8 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
 
     func start(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         if (device != nil || captureSession != nil) {
-            result(FlutterError(code: "MobileScanner",
-                                message: "Called start() while already started!",
+            result(FlutterError(code: MobileScannerErrorCodes.ALREADY_STARTED_ERROR,
+                                message: MobileScannerErrorCodes.ALREADY_STARTED_ERROR_MESSAGE,
                                 details: nil))
             return
         }
@@ -294,8 +298,8 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
         }
         
         if (device == nil) {
-            result(FlutterError(code: "MobileScanner",
-                                message: "No camera found or failed to open camera!",
+            result(FlutterError(code: MobileScannerErrorCodes.NO_CAMERA_ERROR,
+                                message: MobileScannerErrorCodes.NO_CAMERA_ERROR_MESSAGE,
                                 details: nil))
             return
         }
@@ -313,7 +317,9 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
             let input = try AVCaptureDeviceInput(device: device)
             captureSession!.addInput(input)
         } catch {
-            result(FlutterError(code: "MobileScanner", message: error.localizedDescription, details: nil))
+            result(FlutterError(
+                code: MobileScannerErrorCodes.CAMERA_ERROR,
+                message: error.localizedDescription, details: nil))
             return
         }
         captureSession!.sessionPreset = AVCaptureSession.Preset.photo
@@ -476,8 +482,9 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
                     
                 if error != nil {
                     DispatchQueue.main.async {
-                        // TODO: fix error code
-                        result(FlutterError(code: "MobileScanner", message: error?.localizedDescription, details: nil))
+                        result(FlutterError(
+                            code: MobileScannerErrorCodes.BARCODE_ERROR,
+                            message: error?.localizedDescription, details: nil))
                     }
                     return
                 }
@@ -502,10 +509,11 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
             }
             
             try imageRequestHandler.perform([barcodeRequest])
-        } catch let e {
-            // TODO: fix error code
+        } catch let error {
             DispatchQueue.main.async {
-                result(FlutterError(code: "MobileScanner", message: e.localizedDescription, details: nil))
+                result(FlutterError(
+                    code: MobileScannerErrorCodes.BARCODE_ERROR,
+                    message: error.localizedDescription, details: nil))
             }
         }
     }