+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 ## NEXT
 
+Improvements:
 * [MacOS] Added the corners and size information to barcode results.
 * [MacOS] Added support for `analyzeImage`.
 * [MacOS] Added a Privacy Manifest.
 * [web] Added the size information to barcode results.
 * Added support for barcode formats to image analysis.
+* Updated the scanner to report any scanning errors that were encountered during processing.
+* Introduced a new getter `hasCameraPermission` for the `MobileScannerState`.
+* Fixed a bug in the lifecycle handling sample. Now instead of checking `isInitialized`,
+the sample recommends using `hasCameraPermission`, which also guards against camera permission errors.
+
+Bugs fixed:
+* Fixed a bug that would cause the scanner to emit an error when it was already started. Now it ignores any calls to start while it is starting.
+* [MacOS] Fixed a bug that prevented the `anaylzeImage()` sample from working properly.
 
 ## 5.2.3